[AN-4681] Refactors triggering of take over screen for usernames Ticket: https://wearezeta.atlassian.net/browse/AN-4681
So that it only updates the user that calls it.
