+++ app/src/main/java/com/waz/zclient/controllers/usernames/IUsernamesController.java
@@ -37,6 +37,10 @@ public interface IUsernamesController {
 
     void removeUsernamesObserver(UsernamesControllerObserver usernamesControllerObserver);
 
+    void addUsernamesObserverAndUpdate(UsernamesControllerObserver usernamesControllerObserver);
+
+    void closeFirstAssignUsernameScreen();
+
     void logout();
 
     void tearDown();
