+++ app/src/main/java/com/waz/zclient/pages/main/MainTabletFragment.java
@@ -39,7 +39,6 @@ import com.waz.zclient.controllers.confirmation.ConfirmationRequest;
 import com.waz.zclient.controllers.confirmation.IConfirmationController;
 import com.waz.zclient.controllers.navigation.Page;
 import com.waz.zclient.controllers.singleimage.SingleImageObserver;
-import com.waz.zclient.controllers.usernames.UsernamesControllerObserver;
 import com.waz.zclient.core.controllers.tracking.attributes.RangedAttribute;
 import com.waz.zclient.pages.BaseFragment;
 import com.waz.zclient.pages.main.backgroundmain.views.BackgroundFrameLayout;
@@ -53,14 +52,13 @@ import com.waz.zclient.views.menus.ConfirmationMenu;
 
 
 public class MainTabletFragment extends BaseFragment<MainTabletFragment.Container> implements
-        OnBackPressedListener,
-        InAppNotificationFragment.Container,
-        RootFragment.Container,
-        SingleImageObserver,
-        SingleImageFragment.Container,
-        ConfirmationObserver,
-        AccentColorObserver,
-        UsernamesControllerObserver {
+                                                                                   OnBackPressedListener,
+                                                                                   InAppNotificationFragment.Container,
+                                                                                   RootFragment.Container,
+                                                                                   SingleImageObserver,
+                                                                                   SingleImageFragment.Container,
+                                                                                   ConfirmationObserver,
+                                                                                   AccentColorObserver {
 
     public static final String TAG = MainTabletFragment.class.getName();
     private static final String ARG_LOCK_EXPANDED = "ARG_LOCK_EXPANDED";
@@ -124,7 +122,6 @@ public class MainTabletFragment extends BaseFragment<MainTabletFragment.Containe
         getControllerFactory().getAccentColorController().addAccentColorObserver(this);
         getControllerFactory().getAccentColorController().addAccentColorObserver(backgroundLayout);
         getControllerFactory().getBackgroundController().addBackgroundObserver(backgroundLayout);
-        getControllerFactory().getUsernameController().addUsernamesObserver(this);
     }
 
     @Override
@@ -134,7 +131,6 @@ public class MainTabletFragment extends BaseFragment<MainTabletFragment.Containe
         getControllerFactory().getSingleImageController().removeSingleImageObserver(this);
         getControllerFactory().getAccentColorController().removeAccentColorObserver(backgroundLayout);
         getControllerFactory().getBackgroundController().removeBackgroundObserver(backgroundLayout);
-        getControllerFactory().getUsernameController().removeUsernamesObserver(this);
         super.onStop();
     }
 
@@ -263,28 +259,6 @@ public class MainTabletFragment extends BaseFragment<MainTabletFragment.Containe
     @Override
     public void onAccentColorHasChanged(Object sender, int color) {
         confirmationMenu.setButtonColor(color);
-    }
-
-    //////////////////////////////////////////////////////////////////////////////////////////
-    //
-    //  UsernamesObserver
-    //
-    //////////////////////////////////////////////////////////////////////////////////////////
-
-    @Override
-    public void onValidUsernameGenerated(String name, String generatedUsername) {
-        RootFragment fragment = (RootFragment) getChildFragmentManager().findFragmentByTag(RootFragment.TAG);
-        if (fragment != null) {
-            fragment.showFirstTimeAssignUsername(name, generatedUsername);
-        }
-    }
-
-    @Override
-    public void onUsernameAttemptsExhausted(String name) {
-        RootFragment fragment = (RootFragment) getChildFragmentManager().findFragmentByTag(RootFragment.TAG);
-        if (fragment != null) {
-            fragment.showFirstTimeAssignUsername(name, "");
-        }
     }
 
     public interface Container {
