+++ app/src/main/java/com/waz/zclient/pages/main/conversationpager/FirstPageFragment.java
@@ -26,7 +26,6 @@ import android.view.ViewGroup;
 import com.waz.zclient.OnBackPressedListener;
 import com.waz.zclient.R;
 import com.waz.zclient.pages.BaseFragment;
-import com.waz.zclient.pages.main.RootFragment;
 import com.waz.zclient.pages.main.conversationlist.ConfirmationFragment;
 import com.waz.zclient.pages.main.conversationlist.ConversationListManagerFragment;
 import timber.log.Timber;
@@ -124,14 +123,6 @@ public class FirstPageFragment extends BaseFragment<FirstPageFragment.Container>
     @Override
     public void onOpenUrl(String url) {
         getContainer().onOpenUrl(url);
-    }
-
-    @Override
-    public void closeFirstAssignUsernameScreen() {
-        RootFragment fragment = (RootFragment) getChildFragmentManager().findFragmentByTag(RootFragment.TAG);
-        if (fragment != null) {
-            fragment.hideBlankFragment();
-        }
     }
 
     public interface Container {
